Small code fix for when HAPI is being used as server engine in the Trails application to make Primus work with a HAPI server `httpServer.listener` needs to be passed instead of `httpServer` otherwise it won't work.
Small code fix for when HAPI is being used as server engine in the Trails application to make Primus work with a HAPI server `httpServer.listener` needs to be passed instead of `httpServer` otherwise it won't work.

Small code fix for when HAPI is being used as server engine in the Trails application to make Primus work with a HAPI server `httpServer.listener` needs to be passed instead of `httpServer` otherwise it won't work.